move getBrowser to utils.js
