+++ Extensions/combined/src/utils.js
@@ -25,4 +25,18 @@ function numberFormat(numberState) {
   return formatter.format(roundDown(numberState));
 }
 
-export { numberFormat }
+function getBrowser() {
+  if (typeof chrome !== "undefined" && typeof chrome.runtime !== "undefined") {
+    return chrome;
+  } else if (
+    typeof browser !== "undefined" &&
+    typeof browser.runtime !== "undefined"
+  ) {
+    return browser;
+  } else {
+    console.log("browser is not supported");
+    return false;
+  }
+}
+
+export { numberFormat, getBrowser }
